Seperate functionalities, Add eventlistener to refresh button and form
