+++ src/index.html
@@ -23,7 +23,7 @@
     <div id="scoresContent">
       <div id="recentScoresHeader" class="d-flex mb-5 justify-content-between">
         <h2>Recent scores</h2>
-        <button type="button">Refresh</button>
+        <button type="button" id="refreshButton">Refresh</button>
       </div>
 
       <table class="table table-striped table-borderless">
@@ -33,16 +33,19 @@
     </div>
 
     <div id="addScoreContent">
-      <form class="d-flex flex-column">
+      <form class="d-flex flex-column" id="newScoreForm">
         <h2 class="mb-5">Add your score</h2>
         <div class="mb-3">
-          <input class="form-control" type="text" placeholder="Your name" name="name">
+          <input class="form-control" type="text" placeholder="Your name" name="user" required>
         </div>
         <div class="mb-3">
-          <input class="form-control" type="text" placeholder="Your score" name="score">
+          <input class="form-control" type="number" min="1" placeholder="Your score" name="score" required>
         </div>
         <div class="align-self-end ">
           <button type="submit"> Submit </button>
+        </div>
+        <div class="errorMessage">
+          <p>Please input a name and a score</p>
         </div>
       </form>
     </div>
