+++ src/style.scss
@@ -41,3 +41,8 @@ button {
 table {
   border: 4px solid black;
 }
+
+.errorMessage {
+  display: none;
+  color: red;
+}
